feat(all): Add capability, data provenance and dependency graph visualization
this commit adds the execution trace to the interpreter to track the data flow and control flow.

every function call is recorded in the execution trace.

It also adds a new visualizer that uses matplotlib to visualize the dependency graph. the graph can
be shown by invoking the 'graph' command in the agent chat after the code is executed.

It updated the policy enforcing mechanism to track the data provenance and control flow,
taking advantage of the execution trace and dependency graph. It modifies the README to reflect
the new features.

Signed-off-by: Jiaxiao Zhou <duibao55328@gmail.com>